Change tag on gp3 sc & remove it from eks.
Using only the PVCName for the volume names on the AWS console is going
to be confusing if we have more than one deployment (multiple volumes
will have the same name); this commit adds the PVCNamespace as a prefix
to the name to have different names (it is only a label, but seems a
good idea).

Additionally we are not going to use _zoned_ storage classes, so we have
removed their installation from the 'clust/eks' subcommand.