fix(apps): autosave config when using envvars
This commit updates the export_vars functions of the application
installers to set a variable if a value has been read from the
environment instead of from the .env file.

If the variable is set we save the new values when installing the app
and, if the installation succeeds, the updated file is committed to git
(if enabled).