fix: if using sops & yaml, encode after sed
The thing is that 'sops' is too clever and removes quotes if it feels
they are not needed and when we replace __VARIABLES__ by booleans that
must be strings (i.e. on annotations) things fail because sops removed
the quotes when the string value was not a reserved word (true/false).