feat: make create_htpasswd_secret_yaml generic
Now the users need to create the htpasswd file with user:pass lines in
advance and the function will generate a secret with all the users and
passwords hashed.

This change allows us to add more than one user to the secret without
problems (will be used for the zot's htpasswd file).